Bucket 3: bundle size -69%, Admin and Kiosk refactored into modules
#110 — Reduce main bundle size
- vite.config.ts: manualChunks for react-vendor, radix-ui, react-query,
  date-utils, icons — main entry chunk 332kB → 104kB (-69%)
- ChecklistsTab: lazy-load ChecklistBuilderModal, ConvertFileModal,
  BuildWithAIModal, ChecklistPreviewModal; dynamic import export-utils

#107 — Admin refactored into focused modules
- src/pages/admin/shared.ts — types, helpers, parseGoogleOpeningHours
- src/pages/admin/SharedUI.tsx — BottomSheet, modals, ConfirmModal
- src/pages/admin/MyLocationTab.tsx — location management tab
- src/pages/admin/AccountTab.tsx — team/billing/account tab
- Admin.tsx reduced from 2478 → ~300 lines; re-exports public API

#106 — Kiosk refactored into focused modules
- src/pages/kiosk/types.ts — interfaces (Question, KioskChecklist, etc.)
- src/pages/kiosk/utils.ts — pure helpers + UNANSWERED_SENTINEL
- src/pages/kiosk/hooks.ts — useLiveClock, useInactivityTimer
- src/pages/kiosk/KioskSetupScreen.tsx
- src/pages/kiosk/PinEntryModal.tsx — PinDots + NumberPad + modals
- src/pages/kiosk/ChecklistRunner.tsx — runner + all question inputs
- src/pages/kiosk/CompletionScreen.tsx
- Kiosk.tsx reduced from 2686 → ~400 lines; backward-compat re-exports
