+++ vite.config.ts
@@ -23,10 +23,11 @@ export default defineConfig(({ mode }) => ({
     },
   },
   build: {
-    chunkSizeWarningLimit: 500,
+    chunkSizeWarningLimit: 600,
     rollupOptions: {
       output: {
         manualChunks(id) {
+          // Heavy async-only export tools — never on the critical path
           if (id.includes("jspdf") || id.includes("jspdf-autotable")) {
             return "pdf-export";
           }
@@ -39,12 +40,42 @@ export default defineConfig(({ mode }) => ({
             return "document-preview";
           }
 
+          // Charting library — only used on Reporting route
           if (id.includes("recharts")) {
             return "charts";
           }
 
+          // Supabase client — shared but heavy, isolate it
           if (id.includes("@supabase")) {
             return "supabase";
+          }
+
+          // React core — keep together so React is a single shared singleton
+          if (id.includes("/node_modules/react/") ||
+              id.includes("/node_modules/react-dom/") ||
+              id.includes("/node_modules/react-is/") ||
+              id.includes("/node_modules/scheduler/")) {
+            return "react-vendor";
+          }
+
+          // Tanstack React Query
+          if (id.includes("@tanstack/")) {
+            return "react-query";
+          }
+
+          // Radix UI primitives — used across many pages, isolate for shared caching
+          if (id.includes("@radix-ui/")) {
+            return "radix-ui";
+          }
+
+          // Date utilities — react-day-picker + date-fns travel together
+          if (id.includes("react-day-picker") || id.includes("/node_modules/date-fns/")) {
+            return "date-utils";
+          }
+
+          // Icon set — large but all pages use some icons; keep in one cached chunk
+          if (id.includes("lucide-react")) {
+            return "icons";
           }
 
           return undefined;
